BotCompanion/pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "beautifulsoup4>=4.13.4",
    "discord-py>=2.5.2",
    "email-validator>=2.2.0",
    "flask-login>=0.6.3",
    "flask>=3.1.0",
    "flask-sqlalchemy>=3.1.1",
    "google-api-python-client>=2.169.0",
    "gunicorn>=23.0.0",
    "openai>=1.77.0",
    "pillow>=11.2.1",
    "psycopg2-binary>=2.9.10",
    "pynacl>=1.5.0",
    "python-dotenv>=1.1.0",
    "requests>=2.32.3",
    "spotipy>=2.25.1",
    "trafilatura>=2.0.0",
    "youtube-search-python>=1.6.6",
    "yt-dlp>=2025.4.30",
    "werkzeug>=3.1.3",
    "flask-wtf>=1.2.2",
    "sqlalchemy>=2.0.40",
    "requests-oauthlib>=2.0.0",
]
